+++ services/app-web/src/pages/SubmissionSummary/SubmissionSummary.tsx
@@ -17,7 +17,7 @@ import {
     packageName,
     HealthPlanFormDataType,
 } from '../../common-code/healthPlanFormDataType'
-import { makeDateTable } from '../../common-code/data-helpers/makeDocumentDateLookupTable'
+import { makeDateTable } from './makeDocumentDateLookupTable'
 import { base64ToDomain } from '../../common-code/proto/healthPlanFormDataProto'
 import { Loading } from '../../components/Loading'
 import {
